Update files set
