+++ internal/manager/chainlink_node.go
@@ -1,0 +1,17 @@
+package manager
+
+//"context"
+//"fmt"
+//"k8s.io/client-go/kubernetes"
+//"k8s.io/client-go/tools/clientcmd"
+//"k8s.io/client-go/util/homedir"
+//"path/filepath"
+
+//corev1 "k8s.io/api/core/v1"
+//metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+//"k8s.io/apimachinery/pkg/util/intstr"
+//appsv1 "k8s.io/api/apps/v1"
+
+func (m *Manager) ChainlinkNode(network string, env map[string]string) (*NodeInstance, error) {
+	return nil, nil
+}
